Add support for forcing plural link relations.
When there is a one-to-many relationship between a resource and a
particular link relation, a specific representation may return a link
relation that is a json object and other times that same link relation
as a json array. This can be confusing to clients. The setLink method
now allows the user to force a link relation to always be plural if they
so desire.